requirements.txt
gradio==6.0.1
langchain-aws==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
python-dotenv==1.2.1
chromadb==1.3.5
sentence-transformers==5.1.2